this is the greatest accomplishment of my life so far
* load the poster in a hidden `img` attribute so we can figure out
the dimensions for the video player so that horizonal and vertical
videos both render nicely without jumping

The only way to get this in a cross-browser compatabile way is to
use a hidden img element like this. The problem with the things
I would normally use like 'loadedmetadata' and 'canplay' callbacks
on the video element is that none of those fire on Mobile Safari
until _after_ there is user interaction with the player.