SvelteKit example
Since the official [sveltejs/examples](https://github.com/sveltejs/examples) is such a ghost town, I've decided we should just host our sveltekit example ourselves and not bother with a PR, there.

So... here it is! VERY similar to the Remix example, which itself is a minimal take on the classic next.js/examples/with-mux-video